align struct members, not typedefs
aligning a typedef instead of the struct has very strange behavior: the type itself will not be aligned, but a field using the typedef name will have the alignment in other structs. arrays will also not have their elements aligned, but the array as a whole will be aligned.

this is very unintuitive behavior, and we should seek to avoid it. MSVC __declspec(align(N)) annotations work in the more intuitive way, such that annotating a typedef really does annotate the type, so the MSVC struct layout was formerly different; this change should make the MSVC struct layout the same as it is on gcc/clang (by allowing the integer fields of KangarooTwelve_Instance to directly follow finalNode rather than its padding).